+++ src/formatter/format-by-files.test.ts
@@ -1,7 +1,7 @@
 import { ESLint } from 'eslint';
 import stripAnsi from 'strip-ansi';
-import { formatByFiles } from '../../src/formatter/format-by-files.js';
 import { fakeLintResult, fakeLintMessage } from '../test-util/eslint.js';
+import { formatByFiles } from './format-by-files.js';
 
 describe('formatByFiles', () => {
   test('outputs formatted text', () => {
