+++ src/util/array.test.ts
@@ -1,4 +1,4 @@
-import { groupBy, unique } from '../../src/util/array.js';
+import { groupBy, unique } from './array.js';
 
 test('unique', () => {
   expect(unique([0, 1, 1, 2, 1])).toStrictEqual([0, 1, 2]);
